Menu and death implementation, with screens concept
